+++ CHANGELOG.md
@@ -6,6 +6,7 @@ To get the diff for a specific change, go to https://github.com/FriendsOfECMAScr
 To get the diff between two versions, go to https://github.com/FriendsOfECMAScript/BenGorCookies/compare/v0.8.0...v0.9.0
 
 * 0.9.0
+    * Replaced non standard `mousewheel` event by `wheel`.
     * Added Prettier to make CS more consistent.
     * Added Plugin base class that all the plugin must be implemented.
     * Added missing test classes.